+++ pandevice/panorama.py
@@ -53,6 +53,7 @@ class DeviceGroup(VersionedPanObject):
     """
     ROOT = Root.DEVICE
     SUFFIX = ENTRY
+    VSYS_LABEL = 'device-group'
     CHILDTYPES = (
         "firewall.Firewall",
         "objects.AddressObject",
@@ -78,23 +79,126 @@ class DeviceGroup(VersionedPanObject):
 
         self._params = tuple(params)
 
+    @property
+    def vsys(self):
+        return self.name
+
     def devicegroup(self):
         return self
 
     def xpath_vsys(self):
-        if self.name == "shared" or self.name is None:
-            return "/config/shared"
-        else:
-            return "/config/devices/entry[@name='localhost.localdomain']/device-group/entry[@name='%s']" % self.name
+        return self.xpath()
 
-    def _build_xpath(self, root, vsys):
-        if (self.name == "shared" or self.name is None) and root == Root.VSYS:
-            # This is a Vsys object in shared Panorama scope, so proceed normally
-            return super(DeviceGroup, self)._build_xpath(root,
-                self.name or 'shared')
-        else:
-            # This is a member of the device-group, so override to use DeviceGroup root
-            return super(DeviceGroup, self)._build_xpath(self.ROOT, self.name)
+
+class Template(VersionedPanObject):
+    """A panorama template.
+
+    Args:
+        name: Template name
+        description: Description
+        devices (str/list): The list of serial numbers in this template
+        default_vsys: The default vsys in case of a single vsys firewall
+        multi_vsys (bool): (6.1 and lower) Multi virtual systems boolean
+        mode: (6.1 and lower) Can be fips, cc, or normal (default: normal)
+        vpn_disable_mode (bool): (6.1 and lower) VPN disable mode
+
+    """
+    ROOT = Root.DEVICE
+    SUFFIX = ENTRY
+    CHILDTYPES = (
+        "device.Vsys",
+        "device.SystemSettings",
+        "device.PasswordProfile",
+        "device.Administrator",
+        "ha.HighAvailability",
+        "network.EthernetInterface",
+        "network.AggregateInterface",
+        "network.LoopbackInterface",
+        "network.TunnelInterface",
+        "network.VlanInterface",
+        "network.Vlan",
+        "network.VirtualRouter",
+        "network.ManagementProfile",
+        "network.VirtualWire",
+        "network.IkeGateway",
+        "network.IpsecTunnel",
+        "network.IpsecCryptoProfile",
+        "network.IkeCryptoProfile",
+    )
+
+    def _setup(self):
+        # xpaths
+        self._xpaths.add_profile(value='/template')
+
+        # params
+        params = []
+
+        params.append(VersionedParamPath(
+            'description', path='description'))
+        params.append(VersionedParamPath(
+            'devices', vartype='entry', path='devices'))
+        params.append(VersionedParamPath(
+            'default_vsys', exclude=True))
+        params[-1].add_profile(
+            '7.0.0',
+            path='settings/default-vsys')
+        params.append(VersionedParamPath(
+            'multi_vsys', vartype='yesno', path='settings/multi-vsys'))
+        params[-1].add_profile(
+            '7.0.0',
+            exclude=True)
+        params.append(VersionedParamPath(
+            'mode', default='normal', path='settings/operational-mode'))
+        params[-1].add_profile(
+            '7.0.0',
+            exclude=True)
+        params.append(VersionedParamPath(
+            'vpn_disable_mode', vartype='yesno',
+            path='settings/vpn-disable-mode'))
+        params[-1].add_profile(
+            '7.0.0',
+            exclude=True)
+
+        self._params = tuple(params)
+
+    def create_similar(self):
+        raise NotImplementedError('This is not supported for templates')
+
+    def apply_similar(self):
+        raise NotImplementedError('This is not supported for templates')
+
+    def delete_similar(self):
+        raise NotImplementedError('This is not supported for templates')
+
+
+class TemplateStack(VersionedPanObject):
+    """Template stack.
+
+    NOTE:  Template stacks were introduced in PAN-OS 7.0.  Attempting to
+    use this class on PAN-OS 6.1 or earlier will result in an error.
+
+    Args:
+        name: Stack name
+        description: The description
+        templates (str/list): The list of templates in this stack
+
+    """
+    ROOT = Root.DEVICE
+    SUFFIX = ENTRY
+
+    def _setup(self):
+        # xpaths
+        self._xpaths.add_profile(value='/template-stack')
+
+        # params
+        params = []
+
+        params.append(VersionedParamPath(
+            'description', path='description'))
+        params.append(VersionedParamPath(
+            'templates', path='templates', vartype='member'))
+
+        self._params = tuple(params)
 
 
 class Panorama(base.PanDevice):
@@ -115,11 +219,14 @@ class Panorama(base.PanDevice):
     """
     FIREWALL_CLASS = firewall.Firewall
     NAME = "hostname"
+    DEFAULT_VSYS = 'shared'
     CHILDTYPES = (
         "device.Administrator",
         "device.PasswordProfile",
         "firewall.Firewall",
         "panorama.DeviceGroup",
+        "panorama.Template",
+        "panorama.TemplateStack",
     )
 
     def __init__(self,
@@ -154,7 +261,7 @@ class Panorama(base.PanDevice):
         return super(Panorama, self).op(cmd, vsys=None, xml=xml, cmd_xml=cmd_xml, extra_qs=extra_qs, retry_on_peer=retry_on_peer)
 
     def xpath_vsys(self):
-        return "/config/shared"
+        return '/config/shared'
 
     def xpath_panorama(self):
         return "/config/panorama"
@@ -162,7 +269,9 @@ class Panorama(base.PanDevice):
     def panorama(self):
         return self
 
-    def commit_all(self, sync=False, sync_all=True, exception=False, devicegroup=None, serials=(), cmd=None):
+    def commit_all(self, sync=False, sync_all=True, exception=False,
+                   devicegroup=None, serials=(), cmd=None,
+                   description=None, include_template=None):
         """Trigger a commit-all (commit to devices) on Panorama
 
         Args:
@@ -172,6 +281,8 @@ class Panorama(base.PanDevice):
             devicegroup (str): Limit commit-all to a single device-group
             serials (list): Limit commit-all to these serial numbers
             cmd (str): Commit options in XML format
+            description: Commit description
+            include_template (bool): Include template changes in this push
 
         Returns:
             dict: Commit results
@@ -190,6 +301,11 @@ class Panorama(base.PanDevice):
                     d = ET.SubElement(dg_e, "devices")
                     for serial in serials:
                         ET.SubElement(d, "entry", {"name": serial})
+                if description is not None:
+                    ET.SubElement(sp, "description").text = description
+                if include_template is not None:
+                    val = 'yes' if include_template else 'no'
+                    ET.SubElement(sp, "include-template").text = val
             cmd = ET.tostring(e)
         elif isinstance(cmd, pan.commit.PanCommit):
             cmd = cmd.cmd()